requirements.txt
streamlit==1.28.0
pandas==2.2.0
numpy==1.26.0
plotly==5.17.0
scikit-learn==1.3.2
scipy==1.11.4
reportlab==4.0.4
fpdf==1.7.2
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-dateutil==2.8.2
seaborn==0.13.0
matplotlib==3.7.2
